refactor(dataset): split SWE SFT loader into modules
Separate message processing, tokenization, pipeline orchestration, and CLI
code so each concern can evolve without growing a single dataset module.
